Move Controllers to domain packages
arkiv:
- SafController

eux:
- BucController
- SedController

geo:
- LandkodeController
- PostkodeController

pensjon:
- PensjonController

person:
- PersonController
- NavRegistreOppslagController
- AktoerIdHelper

A part of restructuring from Package by Layer to Package by Feature.
