+++ src/main/kotlin/no/nav/eessi/eessifagmodul/arkiv/SafController.kt
@@ -1,4 +1,4 @@
-package no.nav.eessi.eessifagmodul.controllers
+package no.nav.eessi.eessifagmodul.arkiv
 
 import io.swagger.annotations.ApiOperation
 import no.nav.eessi.eessifagmodul.services.saf.SafService
@@ -7,7 +7,10 @@ import no.nav.security.oidc.api.Protected
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.*
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
 import java.util.*
 
 @Protected
